home/jagedo/jagedo/jportal/susrecom/requirements.txt
africastalking==1.2.4
aiohttp==3.8.3
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.2.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
contextlib2==21.6.0
Django==4.1.1
django-admin-interface==0.20.0
django-autocomplete-light==3.9.4
django-autoslug==1.9.8
django-cleanup==6.0.0
django-colorfield==0.7.2
django-crispy-forms==1.14.0
django-crontab==0.7.1
django-datatables-view==1.20.0
django-filter==22.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-generic-m2m==0.3.1
django-haystack==3.2.1
django-model-utils==4.3.1
django-pipeline==2.0.8
django-utils-six==2.0
djangorestframework==3.14.0
frozenlist==1.3.3
idna==3.4
kombu==5.2.4
Markdown==3.4.1
multidict==6.0.4
Pillow==9.2.0
prompt-toolkit==3.0.36
PyMySQL==1.0.2
python-magic==0.4.27
pytz==2022.7
requests==2.28.1
schema==0.7.5
signals==0.0.2
six==1.16.0
slack-sdk==3.19.5
sqlparse==0.4.2
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
yarl==1.8.2
